Fix param check

+++ src/lib/methods/deleteDocument.ts
@@ -24,13 +24,6 @@ export const deleteDocument = function (
 		return callback(err);
 	}
 
-	if (!id || id === String()) {
-		// Throw an error if the data was empty to not stress the servers
-		const err = new Error("No `id` was provided!");
-		if (!callback) return Promise.reject(err);
-		return callback(err);
-	}
-
 	if (typeof id !== "string" && !(id instanceof URL)) {
 		// Throw an error if the data is not a string nor URL
 		const err = new TypeError("Parameter `id` must be a string or an URL!");
@@ -39,6 +32,13 @@ export const deleteDocument = function (
 	}
 
 	const documentId = encodeURIComponent(parseId(id, this.HOSTNAMEREGEX)); // Make the user inputed data encoded so it doesn't break stuff
+
+	if (!documentId) {
+		// Throw an error if the data was empty to not stress the servers
+		const err = new Error("No `id` was provided!");
+		if (!callback) return Promise.reject(err);
+		return callback(err);
+	}
 
 	const opts = prepareRequest({
 		method: "DELETE",
